intial commit to api branch
index.php is now a froncontroller. Added database.php with establishes a
databaseconnection. Added product.php which contains data from the
database. Added create.php which creates new entrys to the database
table. Added read.php which display all data from the database, and
read_one.php which display data from database where id=?. Added 404.php
which is display when an incorrect path is written in the url and
welcome.php which is the landing page.